+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- `linear_operator.utils.errors.NotPSDError` is now catched during training.
+
 ## [1.0.0rc2] - 2026-08-19
 
 - Model: `gpytorch-qr==0.9.0`